Collect UIKit extensions in one file
