+++ Test App/UIKit+Changeset.swift
@@ -1,5 +1,5 @@
 //
-//  UICollectionView+Changeset.swift
+//  UIKit+Changeset.swift
 //  Changeset
 //
 //  Created by Bart Whiteley on 2/12/16.
@@ -11,7 +11,25 @@
 import UIKit
 import Changeset
 
-public extension UICollectionView {
+extension UITableView {
+	
+	/// Performs batch updates on the table view, given the edits of a Changeset, and animates the transition.
+	public func updateWithEdits<T: Equatable> (edits: [Edit<T>], inSection section: Int) {
+		
+		guard !edits.isEmpty else { return }
+		
+		let indexPaths = batchIndexPathsFromEdits(edits, inSection: section)
+		
+		self.beginUpdates()
+		if !indexPaths.deletions.isEmpty { self.deleteRowsAtIndexPaths(indexPaths.deletions, withRowAnimation: .Automatic) }
+		if !indexPaths.insertions.isEmpty { self.insertRowsAtIndexPaths(indexPaths.insertions, withRowAnimation: .Automatic) }
+		if !indexPaths.updates.isEmpty { self.reloadRowsAtIndexPaths(indexPaths.updates, withRowAnimation: .Automatic) }
+		indexPaths.moves.forEach { self.moveRowAtIndexPath($0.from, toIndexPath: $0.to) }
+		self.endUpdates()
+	}
+}
+
+extension UICollectionView {
 	
 	/// Performs batch updates on the table view, given the edits of a Changeset, and animates the transition.
 	public func updateWithEdits<T: Equatable> (edits: [Edit<T>], inSection section: Int, completion: ((Bool) -> Void)? = nil) {
@@ -29,7 +47,7 @@ public extension UICollectionView {
 	}
 }
 
-internal func batchIndexPathsFromEdits<T: Equatable> (edits: [Edit<T>], inSection section: Int) -> (insertions: [NSIndexPath], deletions: [NSIndexPath], moves: [(from: NSIndexPath, to: NSIndexPath)], updates: [NSIndexPath]) {
+private func batchIndexPathsFromEdits<T: Equatable> (edits: [Edit<T>], inSection section: Int) -> (insertions: [NSIndexPath], deletions: [NSIndexPath], moves: [(from: NSIndexPath, to: NSIndexPath)], updates: [NSIndexPath]) {
 	
 	var insertions = [NSIndexPath]()
 	var deletions = [NSIndexPath]()
